fix rounding errors in r2666

+++ baseutils/file_util.h
@@ -62,7 +62,7 @@ BOOL            file_exists(const TCHAR *file_path);
 char *          file_read_all(const TCHAR *file_path, size_t *file_size_out);
 BOOL            write_to_file(const TCHAR *file_path, void *data, size_t data_len);
 #else
-char *          file_read_all(const char *file_path, uint64_t *file_size_out);
+char *          file_read_all(const char *file_path, size_t *file_size_out);
 #endif
 uint64_t        file_size_get(const TCHAR *file_path);
